Rename durability hint, fix jelly for DisableResumeJobProperty, start jelly for DurabilityHintJobProperty
