+++ src/main/resources/org/jenkinsci/plugins/workflow/job/properties/DurabilityHintJobProperty/config-details.jelly
@@ -1,0 +1,48 @@
+<!--
+The MIT License
+
+Copyright (c) 2017, CloudBees, inc.
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-->
+
+<!--
+  Configure build triggers for WorkflowJob
+-->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+    <!-- build triggers config pane -->
+
+    <f:entry title="${%Durability Hint}" field="durabilityHint">
+        <!-- TODO figure out why this fails to show ANY enum values-->
+        <select class="setting-input" name="durabilityHint">
+            Name is: ${descriptor.getPropertyType(instance,durabilityHint).enumConstants}
+            <j:forEach var="it" items="${descriptor.getPropertyType(instance,durabilityHint).enumConstants}">
+                <f:option value="${it.name()}" selected="${instance != null ? it==instance[field] : it.name()==default}">
+                    ${%All Builds}
+                </f:option>
+            </j:forEach>
+        </select>
+
+
+        <!--<f:enum field="durabilityHint" default="FULL_DURABILITY">
+            <strong>cheese and ${it.name}</strong>
+        </f:enum>-->
+    </f:entry>
+</j:jelly>
